# swagger.yaml
openapi: 3.0.3
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:7070
    description: Generated server url
paths:
...
  /versions/candidates:
    get:
      tags:
        - Candidate version Rest API
      summary: Get versions list
      operationId: getVersionsList
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Version'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
    post:
      tags:
        - Candidate version Rest API
      summary: Create new version
      operationId: createNewVersion
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVersionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
...
  /versions/master/:
    get:
      tags:
        - Master version Rest API
      summary: Get master version
      operationId: getVersionDetails
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterVersionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
  /versions/candidates/{versionCandidateId}:
    get:
      tags:
        - Candidate version Rest API
      summary: Get version details by id
      operationId: getVersionDetails_1
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionDetailedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
...
components:
  schemas:
    DetailedErrorResponse:
      type: object
      properties:
        traceId:
          type: string
        code:
          type: string
        details:
          type: string
        localizedMessage:
          type: string
    CreateVersionResponse:
      type: object
      properties:
        id:
          type: string
    CreateVersionRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    MasterVersionResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        author:
          type: string
        latestUpdate:
          type: string
          format: date-time
        published:
          type: boolean
        inspector:
          type: string
        validations:
          type: array
          items:
            $ref: '#/components/schemas/Validation'
    Validation:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - REGULATION_INTEGRITY
            - TEST
        result:
          type: string
          enum:
            - APPROVED
            - REJECTED
        resultDetails:
          type: string
    Version:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    Inspection:
      type: object
      properties:
        name:
          type: string
        inspector:
          type: string
        result:
          type: string
          enum:
            - APPROVED
            - REJECTED
        resultDetails:
          type: string
    VersionDetailedResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        author:
          type: string
        creationDate:
          type: string
          format: date-time
        latestUpdate:
          type: string
          format: date-time
        published:
          type: boolean
        hasConflicts:
          type: boolean
        inspections:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        validations:
          type: array
          items:
            $ref: '#/components/schemas/Validation'